+++ store/rootmulti/store.go
@@ -1,18 +1,25 @@
 package rootmulti
 
 import (
+	"bufio"
+	"compress/zlib"
 	"fmt"
 	"io"
+	"math"
+	"sort"
 	"strings"
 
 	ics23 "github.com/confio/ics23/go"
 	iavltree "github.com/cosmos/iavl"
+	tmiavl "github.com/cosmos/iavl"
+	protoio "github.com/gogo/protobuf/io"
 	"github.com/pkg/errors"
 	abci "github.com/tendermint/tendermint/abci/types"
 	tmmerkle "github.com/tendermint/tendermint/proto/tendermint/crypto/merkle"
 	dbm "github.com/tendermint/tm-db"
 
 	"github.com/cosmos/cosmos-sdk/codec"
+	"github.com/cosmos/cosmos-sdk/snapshots"
 	"github.com/cosmos/cosmos-sdk/store/cachemulti"
 	"github.com/cosmos/cosmos-sdk/store/dbadapter"
 	"github.com/cosmos/cosmos-sdk/store/iavl"
@@ -29,6 +36,11 @@ const (
 	latestVersionKey = "s/latest"
 	pruneHeightsKey  = "s/pruneheights"
 	commitInfoKeyFmt = "s/%d" // s/<version>
+
+	// Do not change chunk size without new snapshot format (must be uniform across nodes)
+	snapshotChunkSize   = uint64(10e6)
+	snapshotBufferSize  = int(snapshotChunkSize)
+	snapshotMaxItemSize = int(64e6) // FIXME SDK has no key/value limit, so we set an arbitrary limit
 )
 
 var cdc = codec.New()
@@ -526,6 +538,229 @@ func parsePath(path string) (storeName string, subpath string, err error) {
 	return storeName, subpath, nil
 }
 
+//---------------------- Snapshotting ------------------
+
+// Snapshot implements snapshots.Snapshotter. The snapshot output for a given format must be
+// identical across nodes such that chunks from different sources fit together. If the output for a
+// given format changes (at the byte level), the snapshot format must be bumped - see
+// TestMultistoreSnapshot_Checksum test.
+func (rs *Store) Snapshot(height uint64, format uint32) (<-chan io.ReadCloser, error) {
+	if format != snapshots.CurrentFormat {
+		return nil, fmt.Errorf("%w %v", snapshots.ErrUnknownFormat, format)
+	}
+	if height == 0 {
+		return nil, errors.New("cannot snapshot height 0")
+	}
+	if height > uint64(rs.LastCommitID().Version) {
+		return nil, fmt.Errorf("cannot snapshot future height %v", height)
+	}
+
+	// Collect stores to snapshot (only IAVL stores are supported)
+	type namedStore struct {
+		*iavl.Store
+		name string
+	}
+	stores := []namedStore{}
+	for key := range rs.stores {
+		switch store := rs.GetCommitKVStore(key).(type) {
+		case *iavl.Store:
+			stores = append(stores, namedStore{name: key.Name(), Store: store})
+		case *transient.Store, *mem.Store:
+			// Non-persisted stores shouldn't be snapshotted
+			continue
+		default:
+			return nil, errors.Errorf("don't know how to snapshot store %q of type %T", key.Name(), store)
+		}
+	}
+	sort.Slice(stores, func(i, j int) bool {
+		return strings.Compare(stores[i].name, stores[j].name) == -1
+	})
+
+	// Spawn goroutine to generate snapshot chunks and pass their io.ReadClosers through a channel
+	ch := make(chan io.ReadCloser)
+	go func() {
+		// Set up a stream pipeline to serialize snapshot nodes:
+		// ExportNode -> delimited Protobuf -> zlib -> buffer -> chunkWriter -> chan io.ReadCloser
+		chunkWriter := snapshots.NewChunkWriter(ch, snapshotChunkSize)
+		defer chunkWriter.Close()
+		bufWriter := bufio.NewWriterSize(chunkWriter, snapshotBufferSize)
+		defer func() {
+			if err := bufWriter.Flush(); err != nil {
+				chunkWriter.CloseWithError(err)
+			}
+		}()
+		zWriter, err := zlib.NewWriterLevel(bufWriter, 7)
+		if err != nil {
+			chunkWriter.CloseWithError(fmt.Errorf("zlib error: %w", err))
+			return
+		}
+		defer func() {
+			if err := zWriter.Close(); err != nil {
+				chunkWriter.CloseWithError(err)
+			}
+		}()
+		protoWriter := protoio.NewDelimitedWriter(zWriter)
+		defer func() {
+			if err := protoWriter.Close(); err != nil {
+				chunkWriter.CloseWithError(err)
+			}
+		}()
+
+		// Export each IAVL store. Stores are serialized as a stream of SnapshotItem Protobuf
+		// messages. The first item contains a SnapshotStore with store metadata (i.e. name),
+		// and the following messages contain a SnapshotNode (i.e. an ExportNode). Store changes
+		// are demarcated by new SnapshotStore items.
+		for _, store := range stores {
+			exporter, err := store.Export(int64(height))
+			if err != nil {
+				chunkWriter.CloseWithError(err)
+				return
+			}
+			defer exporter.Close()
+			err = protoWriter.WriteMsg(&types.SnapshotItem{
+				Item: &types.SnapshotItem_Store{
+					Store: &types.SnapshotStoreItem{
+						Name: store.name,
+					},
+				},
+			})
+			if err != nil {
+				chunkWriter.CloseWithError(err)
+				return
+			}
+
+			for {
+				node, err := exporter.Next()
+				if err == tmiavl.ExportDone {
+					break
+				} else if err != nil {
+					chunkWriter.CloseWithError(err)
+					return
+				}
+				err = protoWriter.WriteMsg(&types.SnapshotItem{
+					Item: &types.SnapshotItem_IAVL{
+						IAVL: &types.SnapshotIAVLItem{
+							Key:     node.Key,
+							Value:   node.Value,
+							Height:  int32(node.Height),
+							Version: node.Version,
+						},
+					},
+				})
+				if err != nil {
+					chunkWriter.CloseWithError(err)
+					return
+				}
+			}
+			exporter.Close()
+		}
+	}()
+
+	return ch, nil
+}
+
+// Restore implements snapshots.Snapshotter.
+func (rs *Store) Restore(height uint64, format uint32, chunks <-chan io.ReadCloser) error {
+	if format != snapshots.CurrentFormat {
+		return fmt.Errorf("%w %v", snapshots.ErrUnknownFormat, format)
+	}
+	if height == 0 {
+		return fmt.Errorf("%w: cannot restore snapshot at height 0", snapshots.ErrInvalidMetadata)
+	}
+	if height > math.MaxInt64 {
+		return fmt.Errorf("%w: snapshot height %v cannot exceed %v", snapshots.ErrInvalidMetadata,
+			height, math.MaxInt64)
+	}
+
+	// Set up a restore stream pipeline
+	// chan io.ReadCloser -> chunkReader -> zlib -> delimited Protobuf -> ExportNode
+	chunkReader := snapshots.NewChunkReader(chunks)
+	defer chunkReader.Close()
+	zReader, err := zlib.NewReader(chunkReader)
+	if err != nil {
+		return fmt.Errorf("zlib error: %w", err)
+	}
+	defer zReader.Close()
+	protoReader := protoio.NewDelimitedReader(zReader, snapshotMaxItemSize)
+	defer protoReader.Close()
+
+	// Import nodes into stores. The first item is expected to be a SnapshotItem containing
+	// a SnapshotStoreItem, telling us which store to import into. The following items will contain
+	// SnapshotNodeItem (i.e. ExportNode) until we reach the next SnapshotStoreItem or EOF.
+	var importer *tmiavl.Importer
+	for {
+		item := &types.SnapshotItem{}
+		err := protoReader.ReadMsg(item)
+		if err == io.EOF {
+			break
+		} else if err != nil {
+			return fmt.Errorf("invalid protobuf message: %w", err)
+		}
+
+		switch item := item.Item.(type) {
+		case *types.SnapshotItem_Store:
+			if importer != nil {
+				err = importer.Commit()
+				if err != nil {
+					return fmt.Errorf("IAVL commit failed: %w", err)
+				}
+				importer.Close()
+			}
+			store, ok := rs.getStoreByName(item.Store.Name).(*iavl.Store)
+			if !ok || store == nil {
+				return fmt.Errorf("cannot import into non-IAVL store %q", item.Store.Name)
+			}
+			importer, err = store.Import(int64(height))
+			if err != nil {
+				return fmt.Errorf("import failed: %w", err)
+			}
+			defer importer.Close()
+
+		case *types.SnapshotItem_IAVL:
+			if importer == nil {
+				return fmt.Errorf("received IAVL node item before store item")
+			}
+			if item.IAVL.Height > math.MaxInt8 {
+				return fmt.Errorf("node height %v cannot exceed %v", item.IAVL.Height, math.MaxInt8)
+			}
+			node := &tmiavl.ExportNode{
+				Key:     item.IAVL.Key,
+				Value:   item.IAVL.Value,
+				Height:  int8(item.IAVL.Height),
+				Version: item.IAVL.Version,
+			}
+			// Protobuf does not differentiate between []byte{} as nil, but fortunately IAVL does
+			// not allow nil keys nor nil values for leaf nodes, so we can always set them to empty.
+			if node.Key == nil {
+				node.Key = []byte{}
+			}
+			if node.Height == 0 && node.Value == nil {
+				node.Value = []byte{}
+			}
+			err := importer.Add(node)
+			if err != nil {
+				return fmt.Errorf("IAVL node import failed: %w", err)
+			}
+
+		default:
+			return fmt.Errorf("unknown snapshot item %T", item)
+		}
+	}
+
+	if importer != nil {
+		err := importer.Commit()
+		if err != nil {
+			return fmt.Errorf("IAVL commit failed: %w", err)
+		}
+		importer.Close()
+	}
+
+	flushMetadata(rs.db, int64(height), rs.buildCommitInfo(int64(height)), []int64{})
+	return rs.LoadLatestVersion()
+}
+
+//----------------------------------------
+// Note: why do we use key and params.key in different places. Seems like there should be only one key used.
 func (rs *Store) loadCommitStoreFromParams(key types.StoreKey, id types.CommitID, params storeParams) (types.CommitKVStore, error) {
 	var db dbm.DB
 
@@ -575,6 +810,25 @@ func (rs *Store) loadCommitStoreFromParams(key types.StoreKey, id types.CommitID
 
 	default:
 		panic(fmt.Sprintf("unrecognized store type %v", params.typ))
+	}
+}
+
+func (rs *Store) buildCommitInfo(version int64) commitInfo {
+	storeInfos := []storeInfo{}
+	for key, store := range rs.stores {
+		if store.GetStoreType() == types.StoreTypeTransient {
+			continue
+		}
+		storeInfos = append(storeInfos, storeInfo{
+			Name: key.Name(),
+			Core: storeCore{
+				CommitID: store.LastCommitID(),
+			},
+		})
+	}
+	return commitInfo{
+		Version:    version,
+		StoreInfos: storeInfos,
 	}
 }